requirements.txt
opencv-python>=4.7.0
mediapipe>=0.10.0
numpy>=1.22.0
pygame>=2.1.0
pillow>=9.0.0 